New HOST SERVER class
sup guys. This version of the project has been uploaded to a new branch. It implements a class called hostServer, which is initialized in the GUI at the same time as the host class. This server class will be waiting for other hosts to connect and when they connect, it handles their request, which can be quit or retr fileName. Both should work, but I haven't done any testing. The retr would only work with text files.

Modifications have been made to the GUI, host and obvs the new class, hostServer.